requirements.txt
requests
numpy
pandas
mlflow
openpyxl
tensorflow
wandb
pyyaml
flask
redis
gunicorn